requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
discord.py==1.0.1
idna==2.8
multidict==4.5.2
Pybooru==4.1.0
pycparser==2.19
PyNaCl==1.3.0
requests==2.21.0
six==1.12.0
urllib3==1.24.3
websockets==6.0
yarl==1.3.0
youtube-dl==2019.4.30
